requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0

# Google Cloud services
google-cloud-texttospeech>=2.16.0

# Enhanced web scraping and content extraction
goose3>=3.1.17
lxml>=4.9.0
html5lib>=1.1

# Data handling and validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
responses>=0.23.0

# Type checking and linting
mypy>=1.5.0
black>=23.7.0
flake8>=6.0.0

# Development utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Date and time utilities
python-dateutil>=2.8.0

# HTTP client with retry capabilities
urllib3>=2.0.0
httpx>=0.24.0
